fix and share util getUnlistedBalance
