+++ carbonmark/components/pages/Portfolio/AssetProject/index.tsx
@@ -12,9 +12,9 @@ import { AssetWithProject } from "lib/actions";
 import { createProjectLink } from "lib/createUrls";
 import { formatToTonnes } from "lib/formatNumbers";
 import { getFeatureFlag } from "lib/getFeatureFlag";
-import { hasListableBalance } from "lib/isListableToken";
 import { LO } from "lib/luckyOrange";
 import { CategoryName } from "lib/types/carbonmark.types";
+import { hasListableBalance } from "lib/utils/listings.utils";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { FC } from "react";
