+++ carbonmark/components/pages/Portfolio/CarbonmarkAssets.tsx
@@ -3,9 +3,9 @@ import { CreateListing } from "components/CreateListing";
 import { SpinnerWithLabel } from "components/SpinnerWithLabel";
 import { Text } from "components/Text";
 import { addProjectsToAssets, AssetWithProject } from "lib/actions";
-import { isListableToken } from "lib/isListableToken";
 import { LO } from "lib/luckyOrange";
 import { User } from "lib/types/carbonmark.types";
+import { isListableToken } from "lib/utils/listings.utils";
 import Link from "next/link";
 import { FC, useEffect, useState } from "react";
 import { AssetProject } from "./AssetProject";
