+++ carbonmark/components/pages/Retire/FromPortfolio/index.tsx
@@ -4,7 +4,7 @@ import { SpinnerWithLabel } from "components/SpinnerWithLabel";
 import { Text } from "components/Text";
 import { useFetchUser } from "hooks/useFetchUser";
 import { addProjectsToAssets, AssetWithProject } from "lib/actions";
-import { isListableToken } from "lib/isListableToken";
+import { isListableToken } from "lib/utils/listings.utils";
 import { FC, useEffect, useState } from "react";
 import { AssetProject } from "./AssetProject";
 import * as styles from "./styles";
